Extract `RealmFeedStore`, `FeedImageCacheObject`, `FeedImageObject` to its own files
