+++ Tests/FeedStoreChallengeTests.swift
@@ -6,78 +6,6 @@ import XCTest
 import FeedStoreChallenge
 import RealmSwift
 
-class FeedImageObject: Object {
-    @objc dynamic var id: String = ""
-    @objc dynamic var feedDescription: String? = ""
-    @objc dynamic var location: String? = ""
-    @objc dynamic var url: String = ""
-
-    var local: LocalFeedImage {
-        return LocalFeedImage(id: UUID(uuidString: id)!, description: feedDescription, location: location, url: URL(string: url)!)
-    }
-
-    convenience init(image: LocalFeedImage) {
-        self.init()
-        id = image.id.uuidString
-        feedDescription = image.description
-        location = image.location
-        url = image.url.absoluteString
-    }
-
-}
-
-class FeedImageCacheObject: Object {
-    @objc dynamic var timestamp = Date()
-    var feedImageObjects = List<FeedImageObject>()
-
-    var feedImage: [LocalFeedImage] {
-        return feedImageObjects.compactMap { $0.local }
-    }
-
-    static func makeCache(_ feed: [FeedImageObject], timestamp: Date) -> FeedImageCacheObject {
-        let cache = FeedImageCacheObject()
-        cache.timestamp = timestamp
-
-        for image in feed {
-            cache.feedImageObjects.append(image)
-        }
-
-        return cache
-    }
-}
-
-class RealmFeedStore: FeedStore {
-
-    let realm: Realm
-
-    init(config: Realm.Configuration) throws {
-        realm = try Realm(configuration: config)
-    }
-
-    func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-        try! realm.write {
-            realm.deleteAll()
-            completion(nil)
-        }
-    }
-
-    func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-        deleteCachedFeed(completion: { _ in })
-        try! realm.write {
-            realm.add(FeedImageCacheObject.makeCache(feed.map { FeedImageObject(image: $0)}, timestamp: timestamp))
-            completion(nil)
-        }
-    }
-
-    func retrieve(completion: @escaping RetrievalCompletion) {
-        if let cache = realm.objects(FeedImageCacheObject.self).first {
-            completion(.found(feed: cache.feedImage, timestamp: cache.timestamp))
-        } else {
-            completion(.empty)
-        }
-    }
-}
-
 class FeedStoreChallengeTests: XCTestCase, FeedStoreSpecs {
 	
     //  ***********************
@@ -167,7 +95,8 @@ class FeedStoreChallengeTests: XCTestCase, FeedStoreSpecs {
 	// - MARK: Helpers
 	
 	private func makeSUT(file: StaticString = #filePath, line: UInt = #line) -> FeedStore {
-        let config = Realm.Configuration(inMemoryIdentifier: String(describing: FeedStoreChallengeTests.self))
+        var config = Realm.Configuration(inMemoryIdentifier: "\(type(of: self))")
+//        config.url = URL(string: "invalid")
         let sut = try! RealmFeedStore(config: config)
 
         trackForMemoryLeaks(sut: sut,file: file, line: line)
@@ -193,21 +122,21 @@ extension FeedStoreChallengeTests {
 //
 //  ***********************
 
-//extension FeedStoreChallengeTests: FailableRetrieveFeedStoreSpecs {
+extension FeedStoreChallengeTests: FailableRetrieveFeedStoreSpecs {
+
+	func test_retrieve_deliversFailureOnRetrievalError() {
+//		let sut = makeSUT()
 //
-//	func test_retrieve_deliversFailureOnRetrievalError() {
-////		let sut = makeSUT()
-////
-////		assertThatRetrieveDeliversFailureOnRetrievalError(on: sut)
-//	}
+//		assertThatRetrieveDeliversFailureOnRetrievalError(on: sut)
+	}
+
+	func test_retrieve_hasNoSideEffectsOnFailure() {
+//		let sut = makeSUT()
 //
-//	func test_retrieve_hasNoSideEffectsOnFailure() {
-////		let sut = makeSUT()
-////
-////		assertThatRetrieveHasNoSideEffectsOnFailure(on: sut)
-//	}
-//
-//}
+//		assertThatRetrieveHasNoSideEffectsOnFailure(on: sut)
+	}
+
+}
 
 //extension FeedStoreChallengeTests: FailableInsertFeedStoreSpecs {
 //
